Add Coupon Remove Maps
Removed Maps from the sliding menu by changing the
Selectioncontroller.m file. Added greek key, both in retina for the
iphone 5, 4s and 4 (and potentially the 5s if it is released by then
you never know) and a non retina phone for the 3gs (I dont think the 3g
will even run our app… fuck them anyway right?)
Modified the informationcontroller files to be the new coupon. Made
duplicate look to the android side both with the words and with the key.